Resolved conflicts I didn't initially see.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc2007/collada@27384 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/terrain/terrain2/threadeddatafeeder.cpp
@@ -92,8 +92,8 @@ public:
 
     if (!material) return;
 
-    if (material->GetWidth () != data->materialMapWidth ||
-        material->GetHeight () != data->materialMapHeight)
+    if ((unsigned int)material->GetWidth () != data->materialMapWidth ||
+        (unsigned int)material->GetHeight () != data->materialMapHeight)
     {
       material = csImageManipulate::Rescale (material,
         data->materialMapWidth, data->materialMapHeight);
